1. 小程序支持退款-模拟的，因为没有对接第三方支付
演唱会管理页面调整
1. 创建新的演唱会或者编辑后会根据场次和票型自动创建库存数据，初始化为0
2. 增加库存管理按钮
3. 删除演唱会时会同时删除场次票型库存，点击删除按钮后会弹出一个提示框
4. 去掉库存的 上架下架状态
5. 去掉删除的验证逻辑， 比如有库存时不能删除演唱会

+++ vue-element-admin/src/views/concert/list.vue
@@ -44,13 +44,17 @@
         </template>
       </el-table-column>
 
-      <el-table-column align="center" label="操作" width="240">
+      <el-table-column align="center" label="操作" width="300">
         <template slot-scope="scope">
           <el-button type="primary" size="mini" @click="handleUpdate(scope.row)">
             编辑
           </el-button>
-          <el-button size="mini" type="danger" @click="handleDelete(scope.row.id,$index)">
-            删除
+          <el-button type="primary" size="mini" icon="el-icon-edit">
+            <router-link :to="'/concert/inventory/'+scope.row.id">
+              库存管理
+            </router-link>
+          </el-button>
+          <el-button slot="reference" size="mini" type="danger" @click="handleDelete(scope.row.id,$index)">删除
           </el-button>
         </template>
       </el-table-column>
@@ -241,7 +245,9 @@ export default {
       })
     },
     handleUpdate(row) {
-      this.temp = Object.assign({}, row) // copy obj
+      this.resetTemp()
+      this.temp = JSON.parse(JSON.stringify(row)) // copy obj
+
       this.dialogStatus = 'update'
       this.dialogFormVisible = true
       this.$nextTick(() => {
@@ -266,15 +272,23 @@ export default {
       })
     },
     handleDelete(id, index) {
-      deleteItem(id).then(() => {
-        this.handleFilter()
-        this.dialogFormVisible = false
-        this.$notify({
-          title: '成功',
-          message: '删除成功',
-          type: 'success',
-          duration: 2000
-        })
+      this.$alert('删除后对应的场次、票型及库存也会一并删除，确认继续么？', '提示', {
+        confirmButtonText: '确定',
+        callback: action => {
+          if (action !== 'confirm') {
+            return
+          }
+          deleteItem(id).then(() => {
+            this.handleFilter()
+            this.dialogFormVisible = false
+            this.$notify({
+              title: '成功',
+              message: '删除成功',
+              type: 'success',
+              duration: 2000
+            })
+          })
+        }
       })
     },
     removeSession(item) {
